chore: move jans-auth base configuration to jans-core-service
