--- jans-auth-server/persistence-model/src/main/java/io/jans/as/persistence/model/base/Entry.java
@@ -1,50 +1,0 @@
-/*
- * Janssen Project software is available under the Apache License (2004). See http://www.apache.org/licenses/ for full text.
- *
- * Copyright (c) 2020, Janssen Project
- */
-
-package io.jans.as.persistence.model.base;
-
-import io.jans.orm.annotation.DN;
-
-/**
- * Provides DN attribute
- *
- * @author Yuriy Movchan Date: 10.07.2010
- */
-public class Entry {
-
-    @DN
-    private String dn;
-
-    public Entry() {
-    }
-
-    public Entry(String dn) {
-        super();
-        this.dn = dn;
-    }
-
-    public String getDn() {
-        return dn;
-    }
-
-    public void setDn(String dn) {
-        this.dn = dn;
-    }
-
-    public String getBaseDn() {
-        return dn;
-    }
-
-    public void setBaseDn(String dn) {
-        this.dn = dn;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("Entry [dn=%s]", dn);
-    }
-
-}
