+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/AppInitializer.java
@@ -21,8 +21,6 @@ import com.google.common.collect.Lists;
 import io.jans.as.common.service.common.ApplicationFactory;
 import io.jans.as.model.common.FeatureFlagType;
 import io.jans.as.model.configuration.AppConfiguration;
-import io.jans.as.persistence.model.configuration.GluuConfiguration;
-import io.jans.as.persistence.model.configuration.IDPAuthConf;
 import io.jans.as.server.model.auth.AuthenticationMode;
 import io.jans.as.server.model.config.ConfigurationFactory;
 import io.jans.as.server.service.cdi.event.AuthConfigurationEvent;
@@ -34,6 +32,8 @@ import io.jans.as.server.service.logger.LoggerService;
 import io.jans.as.server.service.stat.StatService;
 import io.jans.as.server.service.stat.StatTimer;
 import io.jans.as.server.service.status.ldap.LdapStatusTimer;
+import io.jans.config.GluuConfiguration;
+import io.jans.config.IDPAuthConf;
 import io.jans.exception.ConfigurationException;
 import io.jans.model.AuthenticationScriptUsageType;
 import io.jans.model.SimpleProperty;
