+++ jans-config-api/server/src/main/java/io/jans/configapi/service/auth/CouchbaseConfService.java
@@ -8,8 +8,9 @@ package io.jans.configapi.service.auth;
 
 import com.github.fge.jackson.JacksonUtils;
 import com.google.common.collect.Lists;
-import io.jans.as.persistence.model.configuration.GluuConfiguration;
-import io.jans.as.persistence.model.configuration.IDPAuthConf;
+
+import io.jans.config.GluuConfiguration;
+import io.jans.config.IDPAuthConf;
 import io.jans.orm.couchbase.model.CouchbaseConnectionConfiguration;
 import io.jans.service.EncryptionService;
 import io.jans.util.security.StringEncrypter;
