+++ jans-core/service/pom.xml
@@ -72,6 +72,14 @@
 		</dependency>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
+			<artifactId>jans-core-message</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>jans-core-document-store</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
 			<artifactId>jans-core-script</artifactId>
 		</dependency>
 		<dependency>
